services/critic-agent/pyproject.toml
[project]
name = "critic-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "langchain>=0.3.0",
    "langchain-openrouter>=0.2.1",
    "pydantic-settings>=2.6.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
